fixes to JSoup to W3C document conversion to allow for SVG usage
Open HTML to PDF's built-in support for this document tree conversion
breaks SVG support when using the Batik renderer. luckily, more recent
versions of JSoup include their own W3C document tree converter, so we
just use that one instead. and as a bonus, we're no longer using an
ancient version of JSoup.