update how custom font paths are interpreted
i don't actually think it was possible to supply a relative path
with the way this worked before? now it definitely is, which is super
nice for loading fonts from the classpath.

absolute font paths can still be used if needed, but they absolutely
must include a scheme prefix, e.g. 'file:/path/to/font.ttf'